Redesign the Input dialog (#1226)
