+++ src/frontend/qt_sdl/InputConfig/resources/LICENSE.md
@@ -1,0 +1,6 @@
+These vector images are modified from the [Nintendo DS Lite illustration on dimensions.com](https://www.dimensions.com/element/nintendo-ds-lite).
+
+These have been used with the permission of the copyright holders.
+> "We restrict the usage of our drawings and 3D models in commercial software, but as long as it's a free and open source community project, that would be approved. Any reference/backlink to Dimensions.com that could be provided in the developer notes and/or credits for the project would be sufficient for use."
+
+https://www.dimensions.com/legal
